+++ src/components/LandingPage/Sections/GetStartedSection/getstarted-section.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Button from '../../../Button/button';
-import Link from '../../../Link/Link';
+import { Link, Button } from '../../../';
 
 import styles from './getstarted-section.module.scss';
