+++ src/components/index.ts
@@ -1,0 +1,22 @@
+import Link from './Link/link';
+import Button from './Button/button';
+import Header from './Header/header';
+
+import CommunitySection from './LandingPage/Sections/CommunitySection/community-section';
+import FeaturesSection from './LandingPage/Sections/FeaturesSection/features-section';
+import GetStartedSection from './LandingPage/Sections/GetStartedSection/getstarted-section';
+import MainSection from './LandingPage/Sections/MainSection/main-section';
+import PartnerSection from './LandingPage/Sections/PartnerSection/partner-section';
+import SandboxSection from './LandingPage/Sections/SandboxSection/sandbox-section';
+
+export {
+  Link,
+  Button,
+  Header,
+  CommunitySection,
+  FeaturesSection,
+  GetStartedSection,
+  MainSection,
+  PartnerSection,
+  SandboxSection,
+};
